// client/api/model/updateWorkerRequest.ts
/**
 * Smart Power Electric API
 * Smart Power Electric API Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class UpdateWorkerRequest {
    /**
    * Name of the worker
    */
    'name': string | null;
    /**
    * Speciality of the worker
    */
    'speciality': string | null;
    /**
    * Contact of the worker
    */
    'contact': string | null;
    /**
    * Email of the worker
    */
    'address': string | null;
    /**
    * Phone of the worker
    */
    'phone': string | null;
    /**
    * City of the worker
    */
    'socialSecurity': string | null;
    /**
    * State of the worker
    */
    'startDate': string | null;
    /**
    * Zip of the worker
    */
    'endDate': string | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "speciality",
            "baseName": "speciality",
            "type": "string"
        },
        {
            "name": "contact",
            "baseName": "contact",
            "type": "string"
        },
        {
            "name": "address",
            "baseName": "address",
            "type": "string"
        },
        {
            "name": "phone",
            "baseName": "phone",
            "type": "string"
        },
        {
            "name": "socialSecurity",
            "baseName": "socialSecurity",
            "type": "string"
        },
        {
            "name": "startDate",
            "baseName": "startDate",
            "type": "string"
        },
        {
            "name": "endDate",
            "baseName": "endDate",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return UpdateWorkerRequest.attributeTypeMap;
    }
}